[sfm] moved functions inside the lib
