+++ src/software/SfM/main_ConvertSfM_DataFormat.cpp
@@ -5,7 +5,7 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 #include "openMVG/sfm/sfm.hpp"
-#include "openMVG/exif/exif_IO_EasyExif.hpp"
+#include "openMVG/sfm/utils/sfm_data_UID_utils.hpp"
 
 #include "third_party/cmdLine/cmdLine.h"
 #include "third_party/stlplus3/filesystemSimplified/file_system.hpp"
@@ -20,127 +20,6 @@
 
 using namespace openMVG;
 using namespace openMVG::sfm;
-
-/**
- * @brief Update all viewID referenced in the observation of each landmark according 
- * to the provided mapping.
- * 
- * @param[in,out] landmarks The landmarks to update.
- * @param[in] oldIdToNew The mapping between the old ID and the reconmputed UID.
- */
-void updateStructureWithNewUID(Landmarks &landmarks, const std::map<std::size_t, std::size_t> &oldIdToNew)
-{
-  // update the id in the visibility of each 3D point
-  for(auto &iter : landmarks)
-  {
-    Landmark& currentLandmark = iter.second;
-    
-    // the new observations where to copy the existing ones
-    // (needed as the key of the map is the idview)
-    Observations newObservations;
-    
-    for(const auto &iterObs : currentLandmark.obs)
-    {
-      const auto idview = iterObs.first;
-      const Observation &obs = iterObs.second;
-
-      newObservations.emplace(oldIdToNew.at(idview), obs);
-    }
-    
-    assert(currentLandmark.obs.size() == newObservations.size());
-    currentLandmark.obs.swap(newObservations);
-  }  
-}
-
-/**
- * @brief Update all viewID of a list of view by replacing them with the UID.
- * 
- * @param[in,out] views The list of views to update
- * @param[out] oldIdToNew oldIdToNew A map that holds the mapping between the 
- * old ID and the reconmputed UID.
- */
-void updateViewIDs(sfm::Views &views, std::map<std::size_t, std::size_t> &oldIdToNew)
-{
-  // if the views are empty, nothing to be done. 
-  if(views.empty())
-    return;
-  
-  Views newViews;
-
-  for(auto const &iter : views)
-  {
-    const View& currentView = *iter.second.get();
-    const auto &imageName = currentView.s_Img_path;
-    
-    exif::Exif_IO_EasyExif exifReader(imageName);
-
-    // compute the view UID
-    const std::size_t uid = exif::computeUID(exifReader, imageName);
-
-    // update the mapping
-    assert(oldIdToNew.count(currentView.id_view) == 0);
-    oldIdToNew.emplace(currentView.id_view, uid);
-    
-    // add the view to the new map using the uid as key and change the id
-    assert(newViews.count(uid) == 0);
-    newViews.emplace(uid, iter.second);
-    newViews[uid]->id_view = uid;
-  }
-  
-  assert(newViews.size() == views.size());
-  views.swap(newViews);
-}
-
-/**
- * @brief It perform a sanity check on a list of Landmarks and check if the viewIDs
- * in the observations of each landmark has a corresponding ID in the list of views. 
- * @param[in] landmarks A list of landmarks.
- * @param[in] views A list of views.
- */
-void sanityCheckLandmarks(const sfm::Landmarks &landmarks, const sfm::Views &views)
-{
-  for(const auto &iter : landmarks)
-  {
-    const Landmark& currentLandmark = iter.second;
-    for(const auto &iterObs : currentLandmark.obs)
-    {
-      const auto idview = iterObs.first;
-      const Observation &obs = iterObs.second;
-
-      // there must be a view with that id (in the map) and the view must have 
-      // the same id (the member)
-      assert(views.count(idview) == 1);
-      assert(views.at(idview)->id_view == idview);
-    }
-  }  
-}
-
-/**
- * @brief Recompute the UID from the metadata of the original input images and 
- * modify the ID if it's not the same.
- * 
- * @param[in,out] sfmdata The sfmdata scene for which to recompute the UID.
- * @param[out] oldIdToNew A map that holds the mapping between the old ID and the 
- * reconmputed UID.
- * @param[in] sanityCheck Enable a sanity check at the end to assure that the 
- * observations of 3D points and the control points have been correctly updated.
- */
-void regenerateUID(sfm::SfM_Data &sfmdata, std::map<std::size_t, std::size_t> &oldIdToNew, bool sanityCheck = false)
-{
-  // if the views are empty, nothing to be done. 
-  if(sfmdata.GetViews().empty())
-    return;
-  
-  updateViewIDs(sfmdata.views, oldIdToNew);
-  
-  if(!sanityCheck)
-    return;
-  
-  sanityCheckLandmarks(sfmdata.GetLandmarks(), sfmdata.GetViews());
-  
-  sanityCheckLandmarks(sfmdata.GetControl_Points(), sfmdata.GetViews());
-  
-}
 
 // Convert from a SfM_Data format to another
 int main(int argc, char **argv)
